server/pyproject.toml
[tool.poetry]
name = "publication-manager"
version = "0.1.0"
description = ""
authors = ["Sairam M <sairam.m.2005.65@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.1"
uvicorn = "^0.30.6"
python-multipart = "^0.0.9"
celery = "^5.4.0"
redis = "^5.0.8"
supabase = {git = "https://github.com/supabase/supabase-py.git"}
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
